Refactor cli to use subcommands. Update docs. Fixed left edge for partial check.
