+++ .gitattributes
@@ -4,3 +4,4 @@ test/input/chr21_cens.fa.out filter=lfs diff=lfs merge=lfs -text
 test/input/chr21_cens_partials.fa.out filter=lfs diff=lfs merge=lfs -text
 test/input/chr21_chr13_cens_mismap.fa.out filter=lfs diff=lfs merge=lfs -text
 test/input/chr22_cens.fa.out filter=lfs diff=lfs merge=lfs -text
+test/input/chr4_cens_partials.fa.out filter=lfs diff=lfs merge=lfs -text
