+++ test/expected/correct_chr21_cens_partials.tsv
@@ -25,7 +25,7 @@ K1463_2212_chr21_pat-0000578:33292290-37813886	K1463_2212_chr21_pat-0000578:3329
 downsampled_K1463_2209_chr21_pat-0000413:33292621-38201667	downsampled_K1463_2209_chr21_pat-0000413:33292621-38201667	rev	false
 K1463_2214_chr21_contig-0001285:1440556-4112977	K1463_2214_chr21_contig-0001285:1440556-4112977	fwd	true
 downsampled_K1463_2188_chr21_pat-0000683:2025058-4069266	downsampled_K1463_2188_chr21_pat-0000683:2025058-4069266	rev	true
-K1463_2214_chr21_contig-0000195:2-2407367	K1463_2214_chr21_contig-0000195:2-2407367	fwd	true
+K1463_2214_chr21_contig-0000195:2-2407367	K1463_2214_chr21_contig-0000195:2-2407367	fwd	false
 K1463_2298_chr21_pat-0000575:737317-5203113	K1463_2298_chr21_pat-0000575:737317-5203113	fwd	false
 K1463_2215_chr21_mat-0000007:94350933-96395145	K1463_2215_chr21_mat-0000007:94350933-96395145	fwd	true
 chm13_chr21:7700001-11850000	chm13_chr21:7700001-11850000	fwd	false
