+++ test/expected/correct_chr4_cens_partials.tsv
@@ -1,0 +1,1 @@
+HG03452_chr4_h1tg000048l#1-130418273:2-2292603	HG03452_chr4_h1tg000048l#1-130418273:2-2292603	fwd	true
