+++ test/input/chr4_cens_partials.fa.out
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:f92e4ce7dd699b7309640c78dd4643a0d79a77feb11a82270a91a3b24fc222cd
+size 105892
